refactor: move a network banner in the widgets dir
- move history scroll code in the Bootstrap
