+++ packages/frontend/src/app/index.css
@@ -83,4 +83,7 @@
     --mnw-color-10: #272729;
     --mnw-color-11: #72727a;
     --mnw-color-12: #787F85;
+    --mnw-color-13: #f6c98e;
+    --mnw-color-14: #452500;
+    --mnw-color-15: #A15600;
 }
