+++ packages/frontend/src/app/ui/Layout/index.tsx
@@ -1,16 +1,19 @@
-import React, { FC } from 'react';
+import React, { FC, ReactNode } from 'react';
 
 import Footer from '../../../components/common/Footer';
 import User from '../../../widgets/User';
+import NetworkBanner from '../../../widgets/NetworkBanner';
 import { StyledLayout, StyledTop, StyledUserWrapper } from './ui';
 
 type MainProps = {
-    children: React.ReactChildren;
+    children: ReactNode;
 };
 
 const Layout: FC<MainProps> = ({ children }) => {
     return (
         <StyledLayout>
+            <NetworkBanner />
+
             <StyledTop>
                 <StyledUserWrapper>
                     <User />
