--- packages/frontend/src/utils/ScrollToTop.js
@@ -1,17 +1,0 @@
-import { useEffect } from 'react';
-import { withRouter } from 'react-router-dom';
-
-function ScrollToTop({ history }) {
-    useEffect(() => {
-        const unlisten = history.listen(() => {
-            window.scrollTo(0, 0);
-        });
-        return () => {
-            unlisten();
-        };
-    }, []);
-
-    return (null);
-}
-
-export default withRouter(ScrollToTop);
